db_transfer/constants.py
import os
from dotenv import load_dotenv

load_dotenv()  # take environment variables from .env.

POSTGRES_PASSWORD=os.getenv("POSTGRES_PASSWORD")
POSTGRES_PORT=os.getenv("POSTGRES_PORT")
POSTGRES_HOST=os.getenv("POSTGRES_HOST")

